[lsp][extract] Only keep `tparams_rev` that are used by the synthesized type
Summary:
This diff continues the goal of adding generic type parameters that becomes unbounded after refactor. Currently, the `tparams_rev` we collect are all the type parameters that are in scope for the expression. We want only the ones are used so that we don't need to unnecessarily add type parameters to generated `newFunction()` definition.

For example, for the following function

```
function <A, B>foo(v: B) { const a = v; }
```

previously it will report `[A, B]` as `tparams_rev`, but after this diff it will only report `[B]` as `tparams_rev`. The more accurate list allows us to generate `function newFunction<B>(v: B) {...}` instead of `function newFunction<A, B>(v: B) {...}`.

It will also take account of the names used in constraints of the type parameter. For example, the used names of `d` for the following case will be correctly classified as `D, C, B`:

```
function test<A, B, C:B, D:B=C>(d: D) {}
```

Reviewed By: vrama628

Differential Revision: D29724478

fbshipit-source-id: ec7761d1672404ae237fd5c43786d34d2574b53b