[lsp][extract][easy] Generalize function parameter names
Summary:
The extract to constant refactor should also support extracting to multiple scopes, thus also having the undefined variable after move problem. Therefore, The undefined variable analysis will also be useful for extract to constant.

This diff renames the parameter so that it's obvious that they can work for both statements and expressions.

Reviewed By: vrama628

Differential Revision: D29779836

fbshipit-source-id: 7cbb7ae0761f792a300b4396c2061b5e4f62ec4a