Changes for -safe-string compatibility
Summary:
A few test targets relied on the -unsafe-string option, where the types `string`
and `bytes` can be implicitly cast back and forth. This is not supported in
OCaml 4.10.

Happily, all but one case was Unix.read/single_read, which do not mutate the
string argument, and are fixed by using the appropriate read_substring function.

In the last case, we read an exclamation mark into `bytes`, then compare with a
string. I added the explicit conversion here from bytes to string. In practice,
we could have used the unsafe conversion, since we are transferring ownership,
but this is a test file so we don't care about peformance.

Reviewed By: jbrown215

Differential Revision: D29775628

fbshipit-source-id: 24b6664f73b01529b97a3d32dd3ceb16d08e5d85