[lsp][extract] Collect a single selected expression
Summary:
This diff starts to collect a single expression selected by the user, as the first step to implement "refactor to constant" refactor. To improve performance, this is done in the same AST pass that collects extracted statements.

The notable catch is that we can only support exact expression now. The current infra can't really support allowing selecting expression with extra whitespaces before and after, since we don't have a way to tell from the AST locs that we selected other parts of statements that do not belong to expression.

For example:

```
const a      = 1;
           ^^^^^^
        // can't tell from AST loc and selected range that
        // `=` is also selected without using raw string search
```

Reviewed By: gkz

Differential Revision: D29700973

fbshipit-source-id: 3bb356b422696b5c9b2b6ba864efd721aa42426f