Load provider locations for havocing from bindings rather than dynamically
Summary: Currently, write locations for havocing in the SSA builder are generated when we see a write to a variable. This means that if we havoc before we've seen any write, we don't have any locs to use to decide if the variable is havocable or not. However, those binding locs are available from the `Bindings.t` value used to generate the SSA env in the first place. This diff makes it use those instead, so that we can properly not havoc a variable when the havoc is caused before we've seen any writes.

Reviewed By: jbrown215

Differential Revision: D29839902

fbshipit-source-id: 35370884f4b29dea00c9f3930a80de60d22bfc15