Add LTI codemod for annotating top level functions
Summary:
This diff adds the basic LTI codemod. It handles top level functions and arrow functions. I'll add more specific cases in follow up diffs as we find them. I've flagged it as "experimental" in the flow command itself to indicate that it will go through improvement

This follows the same basic pattern as the annotate exports codemod. First, the locations of errors are found and the inferred type for each location is stored in a map. It then traverses the AST looking for locations with errors. If it finds one, it adds the annotated type and verifies it (e.g. that it is not too big). Since it is such a similar codemod, I factored out shared code into the `codemod_annotator` base class.

See test cases for what the codemod currently handles.

Reviewed By: jbrown215

Differential Revision: D29626056

fbshipit-source-id: 605c4d6cb92b59d93cb1932841e3c4636bb6eca0