Don't include uninitialized when havocing on scope entry
Summary: For soundness we assume that inside a function, any external var can be uninitialized. However, this is realistically a nonstarter for implementing the environment, because this means that virtually any use of a value at a higher scope is possibly undefined. As a compromise to allow soundness *up to* undefined, this diff removes uninitialized from the writes to variables when entering child scopes.

Reviewed By: jbrown215

Differential Revision: D29840238

fbshipit-source-id: 0e2f6354a31def16352084ebde05a811a5a361b1