[lsp][extract] Includes both functions and methods as insertion points for extract from expression refactor
Summary:
For the extract from statements refactors, we do not include methods as valid insertion points to insert inner functions, since there is already a better, less confusing alternative available: extract to methods. (This also matches TypeScript's behavior.)

However, for extract from constant refactors, we have a different situation. In this case, we are not creating a function to be hoisted, but instead we are inserting a `const newLocal = expr` right before the statement that contains the expression. Therefore, it's perfectly fine to perform that inside a method without introducing more nestings. (It also matches TypeScript's behavior.)

Reviewed By: vrama628

Differential Revision: D29826055

fbshipit-source-id: 5b0825e33df157fe115b2abb1c5ec3b86777505b