Fix LTI codemod so missing locations don't crash
Summary:
This diff fixes the issue that occurs when actually running the codemod. From inspecting the sites where the codemod crashes I believe it happens when:
1. The codemod is trying to annotate an untyped paramater (e.g. `function (x) {...}`)
2. The parameter is typed in another file (e.g. `function (x: SomeType) {...}` where `SomeType` is imported)
3. The imported type definition is a polymorphic function (e.g. `export type SomeType = <X>(X) => string`)

The polymorphism here means it falls into the `Empty_TypeDestructorTriggerT` case. Then, the aloc is converted to a loc. However, the loc is from another file that has not been parsed during the codemod so it cannot get a concrete location. This causes the crash.

I fixed this by adding a check to see if the file of the error location is the same as the one being codemodded (i.e. if it is in the `aloc_tables`). If it is not, printing the error will result in an `unknown` location.

Reviewed By: samwgoldman

Differential Revision: D29809250

fbshipit-source-id: c8951e3cad2bd741de539caeac058f37cb1baa95