[lsp][extract] Support extracting to type alias
Summary:
This diff implements support for extracting selected type to a type alias. To match TypeScript's behavior, it will only extract the type to a type alias statement in the closest scope.

Despite the simplicity of the design, the problem of undefined variable still exists: generic type parameters might become out of scope. They are detected and added as generic type parameter of the extracted `NewType`.

Finding undefined type variables after extraction is using the approach similar to adding undefined generic type parameters: we find a set of used type parameters on the type level (`S_used`) and a set of available type parameters (`S_avail`), and `S_used - S_avail` is the set of undefined type parameters. The set `S_used` is generated from type synthesizer by querying the type level type at the same location of the AST level type.

Reviewed By: vrama628

Differential Revision: D29835338

fbshipit-source-id: 603f0025fb942792b5b6f67745e7e73a118c0bb8