[lsp][extract] Collect named insertion points of constant definition for extract to constant
Summary:
We should allow user to extract to constant refactor into various scopes. Therefore, we also need to provide good names for those scopes.

This diff changes the `AstExtractor` to inherit factored out `function_and_method_insertion_point_visitor` so that it can reuses the same code to detect named functions/methods along the path to the extracted expressions. It will put these named function info in a stack of `constant_insertion_points`, starting from the toplevel scope. When we encounter a statement that contains the expression, we update the `statement_loc` field to be the new statement loc.

With this logic, it matches the TypeScript behavior that the titles of the refactors are `Extract to constant in module scope/function/method 'xyz'` and the insertion location of each scope is as close to the expression as possible.

Reviewed By: vrama628

Differential Revision: D29797278

fbshipit-source-id: 17578ee451d4388161986c9a8356bb6a3bc301c5