[lsp][extract] Support extract to constant refactor
Summary:
This diff bases on the work of the previous few diffs to implement the first working version of extract to constant refactor. In summary, it will produce a list of extract to constant refactors to each function scope, as long as doing so won't result in new undefined variables.

More work needs to be done in future diffs, including handling `await` and refactoring to class fields.

Reviewed By: vrama628

Differential Revision: D29802129

fbshipit-source-id: 1a3feaa6afe2156708019512c488d85ebbc86f20