[lsp][extract] Ensure generated new function/method/local/field/type names are unique in the module
Summary:
This diff fixes a long standing correctness issue where the generated name like `newFunction` can conflict with existing names. Although collision is rare, they do and can happen.

This diff collects all the names in the old AST and try to suffix the name with increasing numbers until it is unique.

Reviewed By: vrama628

Differential Revision: D29868486

fbshipit-source-id: 4a75c37de8a40ec10f67be65766d92ccfbed04b3