Fix fuzzy-path C stubs
Summary:
OCaml/C interop is full of footguns

In this case, the alloc_result function could lead to heap corruption, since we neglected to register the stack pointer `v` with the OCaml runtime. The calls to `caml_copy_string` and `caml_copy_double` will allocate in the OCaml heap, which can trigger a GC, which can cause the block pointed to by `v` to move.

To fix, we register `v` using `CAMLlocal` macros. This will register a pointer to the value in the stack as a GC root. If the block pointed to by `v` moves, then the value itself will be updated to the new address.

I also changed `alloc_matcher` for completeness, even though there is no risk of it's `v` moving, since we don't allocate before returning. I think it's better to include the boilerplate.

Reviewed By: mroch

Differential Revision: D29882213

fbshipit-source-id: 32d55696e4384bd37e5d1f21990c49c6112b4ca6