[env-builder] Instead of possibly throwing, env builder catches abnormal control flow and returns it (currently unused)
Summary: Now that the env builder (and by extension, the ssa builder) are on the critical path for flow typechecking, they have to be able to recover from exceptions gracefully. This diff changes their interface to not throw abnormal control flow exceptions, but instead captures the abnormal kind and returns it. Currently this is not used for anything, but may need to be used to properly error on e.g. toplevel returns, etc.

Reviewed By: samwgoldman

Differential Revision: D29856920

fbshipit-source-id: 49f26ccf6ebf3775b90863a14fe984b6330ecba2