[lsp][extract] Avoid entering extract refactor code path when user selection is zero-width
Summary:
As we prepare to launch extract to function/method/constant/field/typealias refactor, we need to prepare for unexpected performance hits. To mitigate some potential negative impacts, we can ensure that we won't unnecessarily trigger the feature when the user clearly doesn't expect refactor code actions to be provides.

One such example is when the user selection is zero-width. It's guaranteed that a zero-width selection will return exactly zero available refactors, but it will still do a useless full AST pass to figure that out. This diff avoids that by returning an empty list without going into the available refactors computation.

Reviewed By: mvitousek

Differential Revision: D29914425

fbshipit-source-id: 600bbf498656d0e136abbea49e038365226126b3