Fix scope/ssa/env analysis for let-bound variables in switch statements
Summary:
Let-bound variables declared inside switch statements were previously unhandled, as documented by a TODO comment in the scope builder tests. This diff implemented them; it's a little surprising how they work, but variables declared in a switch statement are live to the entire switch statement (other than the target), including the case statements: that is,
```switch (x) {
  case variable:
    break;
  case 0:
    let variable = 42;
}
```
here `variable` in the case statement *does* refer to the let-bound in the different case. This diff handles this correctly.

Reviewed By: samwgoldman

Differential Revision: D29901505

fbshipit-source-id: 46eebb7e32bc3c6e790b21159a3c799931ddbfd9