Invariant
Summary:
`invariant(cond)`  introduces the `cond` refinement to subsequent lines. To model this, we push a new refinement scope, traverse the condition, and then pop the refinement scope **without getting rid of the refinement writes**.

We also take care to support the special cases of invariant, which are `invariant(false)` and `invariant`. Both of these expressions are control-flow throws, which introduces additional complexity because the env_builder does not expect expressions to throw. D29235485 fixes opportunities for potential unnecessary throws.

Reviewed By: mvitousek

Differential Revision: D28991831

fbshipit-source-id: 3629d8cfd8f0330c6175748511b9219511a3bb23