stop using jekyll-assets
Summary:
now that we're using webpack to bundle JS and CSS, jekyll-assets isn't doing much. all we're using it for is the `asset` macro which inserts the right HTML tag and outputs a versioned file.

there are two caches that need to be busted: the browser cache, and the CDN's origin cache (the CDN caches it instead of fetching from S3). we previously used a versioned filename to invalidate both, but we also already invalidate the CDN cache on deploy. so using a query param to invalidate the browser cache should be sufficient.

Reviewed By: SamChou19815

Differential Revision: D29947979

fbshipit-source-id: 62354dbcfa30aad0c56f7c2e811d18b87626f39c