remove "--retry-if-init false"
Summary:
`--retry-if-init false` made a command exit if the server existed, but was still initializing.

i can't think of a good reason you'd want this behavior. use `--no-auto-start` if you don't want to start a server if one wasn't running, but once one is starting why wouldn't you wait for it? use `--timeout` if you don't want to wait too long.

this is a legacy setting inherited from hack, which also no longer supports it.

Reviewed By: samwgoldman

Differential Revision: D29970417

fbshipit-source-id: b57cadb04d0f9519ac8cb6f7bc5103627983ca8f