[ast-differ] Collect parents of statement changes
Summary:
So far, the ast-differ does not provide any information that is helpful for deciding the indentation of statement changes. This is not a huge issue in the past, since we are mostly making toplevel changes, but it becomes a problem for refactor code action, since it might insert functions  that will span multiple lines into inner scopes. Without proper indentation, the resulting code will be ugly.

The diff starts to collect parents of statement changes, so that later `Ast_diff_printer` can use this information to make better indentation decisions.

Reviewed By: vrama628

Differential Revision: D29975257

fbshipit-source-id: f45ad7986ddc14b4f3d50c69f233b4a676b90052