fix LSP request IDs
Summary:
when the LSP process forwards a message to/from the server, it prefixes the message's ID with the server's ID so that the id is unique.

for example, if the server ID is `1` and the message id is `2`, the "wrapped id" is `1:#2`.

there was an off-by-one bug in `decode_wrapped`, causing it to decode that as server ID `1` but message id `#2`, and then `int_of_string` errored. including the `#` is wrong.

instead of offset math, used `scanf` instead.

Reviewed By: vrama628

Differential Revision: D30015913

fbshipit-source-id: 89405f08080437ae827d128500b83487eb9ad201