Refactor LTI codemod to calculate errors during AST traversal
Summary: This diff modifies the order of how the LTI codemod infers types. Previously, it built up a map of locs to types at the beginning of the codemod then just looked up the type in the map when it came across an error location during AST traversal. Now, the map is from a loc to an error. When an error location is found during AST traversal, the type is then inferred at that point. This change is necessary to facilitate annotations on implicit `this` types as the location of the error is different from the inferred type.

Reviewed By: jbrown215

Differential Revision: D29855333

fbshipit-source-id: 0e60cec5f4b11bf0e0d0119602c952e5be3490b4