do not crash on invariant in conditional or logical expressions
Summary:
`invariant()` and `invariant(false, ...)` introduce opportunities for *expressions* (instead of statements) to introduce abnormal control flow.

There are two kinds of expressions that need to care about this, and these are both kinds of expressions that *conditionally* evaluate one of its subexpressions:
1. Conditional expressions
2. Logical expressions

Instead of reporting an `invariant()` expression as an unconditional throw in a conditional or logical expression, we run each branching expression to completion, merge the completion states, and then commit to an abrupt completion if necessary.

Reviewed By: mvitousek

Differential Revision: D29235485

fbshipit-source-id: 19ea75ca1eeb5e8f84d94b0125cc95a9463233b3