switch statements
Summary:
switch statement support! Switch statements do a strict equality check against the target of the switch statement, but they do not introduce sentinel refinements.

breaks/fallthroughs are handled here for free based on the control-flow handling of the ssa_builder. I added special support for switches with default statements, which the ssa_builder did not provide. If a switch has a default case then we _always_ hit a case in the switch statement, otherwise the statement can be skipped completely. This is relevant for abrupt completions

Reviewed By: mvitousek

Differential Revision: D29213582

fbshipit-source-id: 8d1d5f30ab3f2d8ee26ab44dfe159e05de2aa2c2