Track reads, writes, and refinements on globals
Summary:
Important changes are in the env_builder-- tests are mostly unchanged other than adding in references to globals.

Now that we have the full set of unbound names and a way to consume globals, it's time to actually support global variables in the environment.

We do this by initializing an environment filled with globals above the top-level scope and let the implementation of EnvBuilder handle the rest! :)

Reviewed By: mvitousek

Differential Revision: D29498785

fbshipit-source-id: 44c33fc2ce4e7d93ab8b3deb0bdfd361a722ba2a