Track non-bound names in ssa builder
Summary:
This is the first diff that builds towards supporting global names. Here's how that'll work:

Global variables can be modeled as local variables that exist in a scope higher than the top-level scope. The only extra step we need to take for compatibility with the EnvBuilder's implementation is that we must supply the full list of globals ahead of time. This could be achieved in two different ways:
1. Populate the EnvBuilder with the results from the lib def merging step
2. Populate the EnvBuilder with all the missing names in file, some of which might just be unbound variables and not actually globals.

This stack opts for the second strategy-- the set of globals actually used in a file has the potential to be MUCH smaller than the full set of globals. On top of that, not much code actually contains unbound vars because those would be runtime errors.

Since the SSA builder runs before the env_builder and already traverses every variable name in the file, we can just keep track of them there to populate the EnvBuilder's analysis.

Reviewed By: mvitousek

Differential Revision: D29498783

fbshipit-source-id: 1a3500f8afb3579330f089cfca4969d3db783331