another LSP message ID fix
Summary:
follow on to D30015913 (https://github.com/facebook/flow/commit/8ab4bf699fde8b6b38b7e19ab71872dbbd718416): before sending a response to the server, we need to unwrap its ID.

server 2 (e.g. the server has restarted once) sends a request to the LSP process with message ID `1`. the LSP process forwards this to the client as "2:#1".

when the LSP process gets the response back, it needs to forward the response to server 2 with message ID `1` again.

currently, it's sending back `2:#1` instead.

Reviewed By: SamChou19815

Differential Revision: D30030231

fbshipit-source-id: 70b92350f9bdf3e168a2b91857047fab94b5158d