[lsp] add source.addMissingImports command
Summary:
adds a `workspace/executeCommand` command called `source.addMissingImports` that triggers an edit to fix all of the undefined variables in that file which have an unambiguous import (see `Code_action_service.autofix_imports`).

`workspace/executeCommand` is triggered by the client and can't directly make edits. instead, the server has to send the client a `workspace/applyEdit` request out of band. we wait for the `applyEdit` response and then respond to the `executeCommand`, so that if the `applyEdit` fails, the `executeCommand` can fail too.

this is the first server -> client request we do. we do some lsp -> client requests, like for showing status messages or fetching configuration, but have never done server -> lsp -> client as we do here. therefore, there's some new stuff to keep track of response handlers and whatnot.

Changelog: Adds an LSP command for auto-fixing all unambigous missing imports in a file ( `source.addMissingImports`, which takes a [TextDocumentIdentifier](https://microsoft.github.io/language-server-protocol/specifications/specification-current/#textDocumentIdentifier) as its only argument).

Reviewed By: SamChou19815

Differential Revision: D30030059

fbshipit-source-id: 1b66a6fef53cdb57a26a1c0b6de70535818064c7