fix pretty printer issue with breaking on groups
Summary:
the `Group` layout node tells the pretty printer that it should try to fit the contained nodes on one line, but break the outermost group into multiple lines if they don't.

when determining if a group fits on a line, we weren't considering the rest of the line after the group.
```
if (x) { y; }
```
`(x)` is a group and `{ y; }` is a group. when checking `(x)`, we knew we started at column 3 so had 80 - 3 = 77 chars to fit `(x)`.

this is wrong! we also have to fit `{` -- but we don't have to fit `{ y; }`, because _that_ group can break and only leave `{` on the line (see below).

firstly, we now pass a `rest` param to `fits` containing the rest of the layout nodes so that we can continue to look at the rest of the line.

secondly, we also have to keep track of the modes of parent groups. if we're looking at a nested group whose parent is wrapped (`mode = Break`), then we know we're at the end of a line when we see `IfBreak (Newline, _)`. this is how we know that we only have to fit `{`, not `{ y; }`.

before, we wouldn't have wrapped at all. if we did the first part but not the second, we'd get
```
if (
  x
) { y; }
```
but because we try not to wrap the current group and allow groups in `rest` to break, we end up with the desired:

```
if (x) {
  y;
}
```

Reviewed By: pieterv

Differential Revision: D30106748

fbshipit-source-id: 4b8ee044a60b722b9a6cd3779a477e56cb45e10b