Add $Flow$ModuleRef intrinsic to prelude
Summary:
Prelude is supposed to contain definitions for any builtins which are looked up
by the checker, like Promise for async/await. When a module ref prefix is
defined in the .flowconfig, we will look up this builtin definition.

This change makes it possible to use the --no-flowlib option on files that use
module refs. This diff includes an example, removing the flowlib dependency in
the module_ref tests.

Reviewed By: mroch

Differential Revision: D30171738

fbshipit-source-id: ae929a1366f4dce7f67d2f01779049ae7ded7d00