Fix FilterMaybeT when supplied UnionT
Summary:
Previously, if FilterMaybeT got a (non-optimized) UnionT lower bound, it would result in a tvar with lowerbounds rather than a union type. e.g.
```
type S = $NonMaybeType<$Values<{a: number, b: void}>>;
(1: S);
```
results in
```
Cannot cast `1` to `S` because number [1] is incompatible with empty [2]. [incompatible-cast]
```
https://flow.org/try/#0C4TwDgpgBAKlC8UAkA5A9gOwLIEMQCMIZwIAeJANRwBsBXCAZ1IG8cAuKDWgW0ICcANFHwcAbmgCWAEwC+APjkBuAFAAKAIwcYASkVQA9PqgMAFmlrUpw6GgDWy5aEhQA8vgBWCKAH5mAH2UoKAAzNDQOLl4IQWU-GRUnaABlL1RMXAIiElI3d28AOgBtAHJQtGKAXSU1TSgk3QMjU3NLayg7IA

What we can do instead is map the union to another union, using `FilterMaybeT` on each member of the union. `FilterMaybeT` will result in `EmptyT` for null and void, which means things work more the way we expect.

Reviewed By: dsainati1

Differential Revision: D30121669

fbshipit-source-id: ea6b26b80da06c2ddc8b37031906ca42e358e695