fix unnecessary parens around arrow functions
Summary:
when the body of an arrow function is an expression, it's specifically an AssignmentExpression, so it needs parens for things with precedence lower than that (namely, just sequence expressions). otherwise, `() => 1, 2` is `(() => 1), 2`, not `() => (1, 2)`.

however, the implementation caused us to also parenthesize the entire arrow function when used in an AssignmentExpression, as in `x = (() => {})`. this is not ambiguous and does not need parens.

the fix is to make the arrow function's precedence the same as assignment expressions. it must be lower than the next-highest precedence, conditionals, so that we parenthesize `(() => {}) ? 1 : 2`.

relatedly, we were passing the precedence down to the body expression which works coincidentally. since the body expression is an AssignmentExpression, we should use that precedence explicitly.

Reviewed By: panagosg7

Differential Revision: D30277961

fbshipit-source-id: 48c944f0e7a056a6acb8d1a2084aed57a0f0e4ef