store tvar in require_map
Summary:
`Context.require_map` stores a placeholder for the imports of a module.
After the module that provides this import has been resolved the produced
type flows to the placeholder type.

This diff changes require_map to hold type variables instead of any type. In fact
this is actually the case today, but now we also enforce this through the value type
of the map. When a module type gets resolved we use `resolve_id` to resolve
the type variable instead of `flow_t` that we did before.

Motivation for this is to remove the dependency of `connect_require` on `flow_t`
and instead replace with resolve_id, which is easier to implement with annotation
constraints (coming in follow up diffs).

Reviewed By: samwgoldman

Differential Revision: D29382653

fbshipit-source-id: d42335ce6b09a8b52ab818ddbd1bfef949f228ec