Use id loc in RTypeAlias reason
Summary:
Stacked diffs will use this location to fetch documentation. The jsdoc comments
are attached to the identifier AST node, so it's helpful to point to that node
instead.

RTypeAlias was added in D8246064 (https://github.com/facebook/flow/commit/004b8017706309ed6c16ba00a5ea889323d9254c) to make detection of type aliases in the
normalizer more reliable. I tried to re-familiarize myself with that change, and
could find no motivation for the specific choice of loc.

Reviewed By: panagosg7

Differential Revision: D30362535

fbshipit-source-id: cbfb2ff3da47e0b38326ed274fa695be16d8085b