[new-website] Generate Flow type errors for code blocks during markdown compilation
Summary:
This diff switches the strategy of Flow type error generation in JS code blocks from client side generation using `flow.js` to build time generation via Flow binary during markdown compilation stage with a remark plugin.

The remark plugin will look for a code block with "```<any-language> flow-check", and compile that into jsx of the form

```
<FlowCheckCodeBlock flowErrors={flow_errors_from_code}>
  {original_code_inside_code_block}
</FlowCheckCodeBlock>
```

In this way, all the errors are generated at compile time, and we don't have to ship a 3MB flow.js to the end user of the docs site!

Changelog: [Internal]

Reviewed By: mroch

Differential Revision: D30441720

fbshipit-source-id: 49063f9bf2d9dc6c5b180b083470a5763ad2df68