[new-website] Migrate FAQ page to Docusaurus
Summary:
While migrating the FAQ page, I found a bug in the current implementation of remark plugin. The current string concat way produces code blocks with broken indentation, probably due to formatting loss during transpilation. To preserve indentation and formatting, it's better to keep things in AST format.

Therefore, I decided to keep the type of the node with `flow-check` meta string as `code`, but distinguish it from other code block by the language name `flow` and pass the type errors as `meta`. On the client side, I swizzled the CodeBlock component, but redirect to our `FlowCheckCodeBlock` when the lang is `flow` and to the original Docusaurus code block otherwise.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D30457541

fbshipit-source-id: 6664beeb0214e502a5363537a23a81699e8dd438