[enums] Update enums runtime so `.members()` returns a iterator that is iterable, rather than just an iterable
Summary:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Iteration_protocols

Previously, `.members()` returned an iterable, so you could use it where iterables where expected (e.g. `for-of`). However, it is useful to return something that is both an iterator and an iterable. That way you can also use `.next()` manually if you wish. This is what is returned from things like `Map::values()`.

Since for non-mirrored enums `.members()` returns `Map::keys()`, we don't need to change anything. We just need to update the mirrored case (which previously returned an Array, which is an iterable but not an iterator).

After this diff is landed, and the runtime is updated in all applicable places, I will update the Flow type definition of `.members()` as well.

Reviewed By: samwgoldman

Differential Revision: D30318300

fbshipit-source-id: 7bf8ed33b31a60b85cedd44a46d1876f6dcc31b4