indexed writes to unsealed objects store dictionary types
Summary:
If we would like to be able to infer indexer types for unsealed objects, we need to record when an indexer write occurs to these objects. In a special mode, whenever an unsealed object receives an indexed write, we insert the type of the dictionary into the context, in a map keyed by definition location. We can later use this to codemod in indexer types for unsealed objects.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D29767348

fbshipit-source-id: 2136e733c3af8d0630073db1aa85e7e903e2e6a7