Codemod to annotate empty objects with indexers
Summary:
This adds a new codemod to insert indexer annotations on empty objects being used as indexers. If an empty object only receives indexed writes, we will insert an annotation on it. If possible we add it to the declarator for the object, but if there is none we insert a cast expression around the literal.

This codemod runs in a special mode of Flow that allows for the inference of indexers that does not normally run during typechecking.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D30308631

fbshipit-source-id: f8db476b9235a50ce1d47aa78517bde39af7024b