new-merge scaffolding
Summary:
This diff adds a new merge mode, which is expected to replace the way that
type signatures (`Type_sig.t`) are converted to types (`Type.t`). This process
happens primarily in check_service.ml and type_sig_merge.ml, and calls (in
various places) into Flow_js.

Motivation for this change is remove the dependency of this process on Flow_js,
so that we can eventually simplify the code within Flow_js. The merge deals
with type annotations. Type variables that represent these annotations are
immediately resolved to types once processed (0->1 property) as opposed to
gradually accumulate bounds like the general form of the constraints processed
by Flow_js.

For this reason, the new mode does a similar task, but replaces the constraint
solving engine of Flow_js with an annotation-based one (annotation_inference.ml).

This is a large diff but changes here are mostly mechanical:

1/ A new `--new-merge` flag and an `experimental.new_merge` flowconfig
option are added.

2/ Type_sig_merge becomes parametric over APIs `CONS_GEN` and `TVAR` that
encapsulate the main operations or flow_js.ml and tvar.ml that are necessary for
creating types from signatures.

3/ `mk_check_file` of check_service.ml that calls into Type_sig_merge instantiate
`CONS_GEN` and `TVAR` with modules `ConsGen` that calls into Flow_js and `Tvar`.
`ConsGen` actually existed from before, since some of the operations of Type_sig_merge
were already parameterized. This is effectively replicates the current behavior.

Some helper functions of `mk_check_file`, and also `Type_inference_js.add_require_tvars`
are adjusted to ensure that Tvar and Flow_js is not called directly, but instead only
through `ConsGen`. This is to avoid having Unresolved constraints appear during
annotation inference.

4/ Depending on the value of the "new-merge" option the Merge module in
`mk_check_file` is instantiated with the newly added Annotation_inference module
(mentioned above). This module currently includes stubs.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D28496027

fbshipit-source-id: b8424337a3150d90e75b1d760c42dd2b9dac2b5b