add JSON-RPC id to Persistent request/response logs
Summary:
When the server processes an LSP request, it logs "Persistent request" right before it processes it (i.e. after the request has made it through the queue) and "Persistent response" when it finishes processing it. This diff adds the JSON-RPC request id to these logs, to help make request queueing patterns clearer.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D30522354

fbshipit-source-id: 14994a352ec37b4ac9baba11efb547c56311681c