Store workload items as records
Summary: This diff doesn't change anything about the observable behavior or interface of the `WorkdloadStream` data structure, it just rewrites some tuples as records in preparation for adding a `name` field to the record to track names of workload items. For consistency, this diff adds storing of enqueue times for requeued items.

Reviewed By: mroch

Differential Revision: D30560330

fbshipit-source-id: 3f209fa5502b4107301a31adbbbbf1d2ecdab261