Restructure scopes to allow var scopes to contain lexical declarations
Summary:
Currently, the way the scope_api models lexical vs var scopes doesn't really correspond to reality. It models them as entirely distinct: lexical scopes *only* contain lets, and var scopes *only* contain vars (and toplevel function declarations). This model works for a lot of our analyses, and it has the advantage that you can determine whether a variable is let-like or var-like based just on what scope it's in, but it doesn't work for check-time variable resolution, which is what we need for typechecking.

This diff makes scopes operate more like they do in env.ml/scope.ml (the typechecking environment and scopes). In this model, lexical vars can live in var scopes, though of course not vice versa. We accomplish this by changing the hoister so that the non-lexical hoister extends the lexical hoister, and can handle lexical variables similarly when it's looking at a toplevel (the nonlexical hoister still explores more deeply and ignores lexical variables at deeper levels).

We also need to change the scope_builder to not generate the layered var/lex scopes. Previously it would fire off a nonlexical hoister at functions and toplevels, and then a lexical hoister on programs and function bodies. Since the nonlexical hoister collects lexical variables now, we only need one hoister per level. Changes in the scope_builder reflect this.

As mentioned above, some of our analyses distinguish between lets and vars by what scope they're in, and with this change, that's no longer sufficient. This diff also adds a `kind` field to the `Def.t` for variables, generated by the scope builder and determined in the hoister, which represents *per variable* whether it's a let, a var, a function, etc. This `kind` information is going to be extended significatly for typechecking with the new env, but for this diff it suffices to know whether something is redefinable (vars, functions, and parameters) for the purposes of variable coalescing, and if something can be referenced before it's defined (vars and functions) for the purposes of autocomplete.

Finally, this makes some minor changes to the ast_mapper to ensure that the right kind is always selected when inspecting variables. This fixes an unrelated issue in the annotate_lti tests.

Reviewed By: panagosg7

Differential Revision: D30036001

fbshipit-source-id: 4d80b68bbb1f706e0b4a9abadb703e3e68be624a